requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.4.0
cryptography==43.0.1
Deprecated==1.2.14
Django==3.2.25
djangorestframework==3.15.1
exceptiongroup==1.2.2
idna==3.10
importlib-metadata==6.7.0
iniconfig==2.0.0
jwcrypto==1.5.1
packaging==24.0
pluggy==1.2.0
pycparser==2.21
pytest==7.4.4
pytest-django==4.5.2
pytz==2024.2
PyYAML==6.0.1
redis==5.0.8
requests==2.31.0
responses==0.23.3
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
urllib3==2.0.7
wrapt==1.16.0
zipp==3.15.0
